requirements.txt
streamlit  
Pillow  
google-auth-oauthlib  
google-auth-httplib2  
google-api-python-client  
anthropic  